feat: implemented user paged list

+++ dl-api-csharp/Doublelives.Persistence/Repositories/IMenuRepository.cs
@@ -5,9 +5,9 @@ namespace Doublelives.Persistence.Repositories
 {
     public interface IMenuRepository : IRepository<SysMenu>
     {
-        List<string> GetPermissionsByRoleIds(List<long> roleIds, bool activeOnly = true);
+        List<string> GetPermissionsByRoleIds(List<int> roleIds, bool activeOnly = true);
 
-        List<SysMenu> GetTopLevelMenusByRoleIds(List<long> roleIds);
+        List<SysMenu> GetTopLevelMenusByRoleIds(List<int> roleIds);
 
         List<SysMenu> GetSubMenusByParentCode(string id);
     }
